# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation about projekt SasiadWPotrzebie",
        "description": "sasiadwpotrzebie api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "local server"
        },
        {
            "url": "http://localhost:9000",
            "description": "staging server"
        },
        {
            "url": "http://localhost:9000",
            "description": "production server"
        }
    ],
    "paths": {
...
        "/api/ratings": {
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Get all ratings",
                "description": "Fetches a list of all ratings.",
                "operationId": "bc95a944de90b669c26811f0f49abfa6",
                "responses": {
                    "200": {
                        "description": "List of all ratings"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Create a new rating",
                "description": "Adds a new rating to the system.",
                "operationId": "7714a3e13f3eb872beb0df398aeab4ac",
                "responses": {
                    "201": {
                        "description": "Rating created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/ratings/{id}": {
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Get a specific rating",
                "description": "Fetches the details of a specific rating by its ID.",
                "operationId": "24bab34811eeb4ff6ccf8d876c3f9fd7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rating"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the rating"
                    },
                    "404": {
                        "description": "Rating not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Update a specific rating",
                "description": "Updates the details of a specific rating by its ID.",
                "operationId": "5ea447372972cb53d62dbf9d02ae7e7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rating to update"
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Rating updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Rating not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Delete a specific rating",
                "description": "Deletes a specific rating by its ID.",
                "operationId": "60c99789b0d1674b91ccf5c8fe37c980",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rating to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rating deleted successfully"
                    },
                    "404": {
                        "description": "Rating not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}